Don't ignore the tests results but rather have a tolerance rate
Historically, we ignored the test results as we don't have the
manpower to investigate the failures on various architectures.

Since most of the test pass, I've added a 1% failure tolerance
on everything except x86_64, so we do no longer need to ignore the
results. This makes it easier to spat that it's no longer a couple
test that fails on some random architecture but rather a more
severe problem.

In the process I had to disable test_decomp on ppc64le, because it
segfaults.