acceptable_failure_rate.patch: Don't use pytest private API
The private API used here was removed in pytest 5.
There is a public API now, pytest.ExitCode.TESTS_FAILED,
however that is not yet available in pytest 4.

We hardcode the value 1 instead, it is unlikely to be changed.
Once pytest 4 is gone, we can switch to the new API.

See https://github.com/pytest-dev/pytest/issues/5125